requirements.txt
streamlit==0.58
tensorflow-cpu==2.3.0
joblib
